Sets up Orb repo with stack, scripts, fourmolu, GitHub Actions
This sets up Orb as a stack project, adds fourmolu as part of the
`build` scripts, establishes a GitHub Actions workflow,  and adds
the common scripts from all of our public libraries: `build`,
`shell`, and `test`.

Currently, the main `Orb` file is a simple "Hello, world!", and
the test suite has been included but not yet implemented.