Broadens request/response body possibilities beyond Fleece
Up to now, we offered three possibilities for response bodies:

- JSON, as encoded by Fleece (with a provided schema)
- A document/file of some sort
- No body content

...and two possibilities for request bodies:

- JSON, as decoded by Fleece (with a provided schema)
- No body content

In order to support JSON decoding and encoding with other
libraries, as well as HTML, XML, CSV, and plain text responses,
we now require the user to provide an argument to each that is a
function: the request decoding function that decodes the request
body (or produces a desired error); and the response encoding
function that takes the `ContentType` (ADT defined by `orb`) and
encodes the response body.

`addResponseBodyXXX` functions now expect these values. They are
expected in the `handlerResponseBodies`, out of the way of the
rest of the handler logic and without requiring the user to pass
it as an argument to `returnXXX`. The `requestBody` similarly asks
for the decoder.

This unties Orb from Fleece, but this commit does not remove it as
a dependency so that we can continue to provide
`Orb.Response.Schemas`.