Adds support for requests with form data request bodies
This adds another constructor for `RequestBody` that handles
`multipart/form-data`. We interpret the form fields as a mapping of
field name to content, be it a data param or a file upload. This
makes it easy for users to search by the field name or ID. A user
provided function that validates that form allows us to respond
with a 422 containing the validation errors in the manner the
user chooses.

This is designed to support fields that can have more than one
value (checkboxes, multi-select, the `multiple` attribute),
allowing for a (non-empty) list of values for any given field name.
We reject forms when a param and a file share the same field name.