Adds initial Open API support
This adds facilities for generating an open api spec based
on the Handler definitions described by an Orb router.
The `provideOpenApi` function is used anywhere in a router
to label a section of routes being included in an open-api
spec. The spec is given a label at the time `provideOpenApi`
is called. Once the router is build, `mkOpenApi` can be called
on the root router to retrieve the open api description of
any labeled APIs contained within it. The retrieved API
descriptions will have the correct full path as described
by the root router passed to `mkOpenApi`.

This also provides a helper functions that can be used by
applications to provide a main with an subcommands to
run the Orb api (`api`) or generate an Open API spec based
on the application routes (`generate-open-api`). Various
helpers are exposed to allow the user to compose their
own main function however they ultimately need to.