Fixes bug in OpenAPI Fleece instance `nullable` method
A previous commit accidentally removed the call to `collectComponents`
in the `nullable` method of the `Fleece` instance for `FleeceOpenApi`,
when it should have replaced it with the descendant components of the
schema being made nullable. We added a test case to ensure the correct
OpenAPI schema is generated for fleece schemas with non-primitive child
schemas.

Co-authored-by: Janus <janus@flipstone.com>