Uses custom schema collision comparison function
We added a function called `isSameSchemaInfo` to replace the `Eq`
instance of `SchemaInfo` that compares the infos while ignoring the
`openApiNullable` field to avoid false positives when including both the
nullable and non-nullable versions of a fleece schema in a router.

We also removed the debugging `ppShow` call that we inadvertently left
in the previous commit.

Co-authored-by: Janus <janus@flipstone.com>