Adds support for header and query params
This adds two record fields to `Handler` with associated types in the
`HasHandler` class for describing the header and query params processed
by a handler. These are described using schemas as defined by
`beeline-params` and Orb provides an interperter for those schemas that
produces an OpenAPI description of those params.

This revamps the handler execution a bit to be more clear while adding
the parsing of the new params while handling a request.