Allows handlers and permission actions to use different monads
This adds a new associated type and method to `PermissionAction` to
allow applications to perform permission checks in a different monad
than the handler monad. In particular, the handler monad may include
evidence of the permission execution (e.g. a user record) in its
context that cannot be present before the permission action is
executed. In that case, a `PermissionAction` can implement
`runPermissionActionHandler` to perform execution of the handler monad
from its own monad.