Improves OpenAPI spec names and types and error reporting
This uses the new `json-fleece` API to improve the names and types
given to OpenAPI specs.

This includes a change that restricts the characters allowed in the name
of an OpenAPI spec. This default is to allow alphanumeric characters, plus
'.' and '_'. The list of allowed characters can be specified both via
the various functions that build OpenAPI specs and the default command
line generation command.

This also improves the error reporting when schemas have conflicts or
bad names to try to provide context of where the offending spec is
contained within the spec definitions so that the developer can more
easily locate the offending specs within the code.